pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mvplaybook"
version = "0.1.0"
description = "Reusable metric validation playbook: stability, sensitivity, drift, robustness, and uncertainty checks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

authors = [
    {name = "Your Name"}
]

dependencies = [
    "pandas>=1.5",
    "numpy>=1.23",
    "scipy>=1.9",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
mvplaybook = "mvplaybook.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
